+++ src/App.tsx
@@ -10,39 +10,11 @@ import userService, { IUser } from './services/userService';
 import { useUserContext } from './UserContext';
 
 const App = () => {
-  const { setUserContext } = useUserContext();
+  const { setUserContext, storeUserSession, clearUserSession } = useUserContext();
   const [user, setUser] = useState<IUser | null>(null);
   const [isLoading, setIsLoading] = useState(true);
 
   const navigate = useNavigate();
-
-  const storeUserSession = (userData: {
-    accessToken: string;
-    refreshToken: string;
-    user: IUser;
-  }) => {
-    const { user, accessToken, refreshToken } = userData;
-    setUserContext({
-      username: user.username,
-      _id: user._id,
-      profileImage: user.profileImage,
-    });
-    setUser(user);
-    localStorage.setItem('userId', user._id);
-    localStorage.setItem('accessToken', accessToken);
-    localStorage.setItem('refreshToken', refreshToken);
-    setIsLoading(false);
-  };
-
-  const clearUserSession = () => {
-    setUserContext(null);
-    setUser(null);
-    localStorage.removeItem('userId');
-    localStorage.removeItem('accessToken');
-    localStorage.removeItem('refreshToken');
-    navigate('/login', { replace: true });
-    setIsLoading(false);
-  };
 
   const handleLoginSuccess = (userData: {
     accessToken: string;
