+++ src/UserContext.tsx
@@ -1,4 +1,6 @@
 import React, { createContext, useContext, useState, ReactNode } from 'react';
+import { useNavigate } from 'react-router-dom';
+import { IUser } from './services/userService';
 
 interface User {
   username: string;
@@ -11,12 +13,15 @@ interface UserContextType {
   setUserContext: (user: User | null) => void;
   updateContextUsername: (username: string) => void;
   updateContextProfileImage: (profileImage: string) => void;
+  storeUserSession: (userData: { accessToken: string; refreshToken: string; user: IUser }) => void;
+  clearUserSession: () => void;
 }
 
 const UserContext = createContext<UserContextType | undefined>(undefined);
 
 export const UserProvider: React.FC<{ children: ReactNode }> = ({ children }) => {
   const [userContext, setUserContext] = useState<User | null>(null);
+  const navigate = useNavigate();
 
   const updateContextUsername = (newUsername: string) => {
     if (userContext) {
@@ -30,9 +35,40 @@ export const UserProvider: React.FC<{ children: ReactNode }> = ({ children }) =>
     }
   };
 
+  const storeUserSession = (userData: {
+    accessToken: string;
+    refreshToken: string;
+    user: IUser;
+  }) => {
+    const { user, accessToken, refreshToken } = userData;
+    setUserContext({
+      username: user.username,
+      _id: user._id,
+      profileImage: user.profileImage,
+    });
+    localStorage.setItem('userId', user._id);
+    localStorage.setItem('accessToken', accessToken);
+    localStorage.setItem('refreshToken', refreshToken);
+  };
+
+  const clearUserSession = () => {
+    setUserContext(null);
+    localStorage.removeItem('userId');
+    localStorage.removeItem('accessToken');
+    localStorage.removeItem('refreshToken');
+    navigate('/login', { replace: true });
+  };
+
   return (
     <UserContext.Provider
-      value={{ userContext, setUserContext, updateContextUsername, updateContextProfileImage }}
+      value={{
+        userContext,
+        setUserContext,
+        updateContextUsername,
+        updateContextProfileImage,
+        storeUserSession,
+        clearUserSession,
+      }}
     >
       {children}
     </UserContext.Provider>
